Cleanup Core API added common header
