+++ ngraph/core/include/openvino/core/node_output.hpp
@@ -37,7 +37,6 @@ public:
     /// \param node A `shared_ptr` to the node for the output handle.
     /// \param index The index of the output.
     ///
-    /// TODO: Make a plan to deprecate this.
     Output(const std::shared_ptr<Node>& node, size_t index);
 
     /// \brief Constructs a Output, referencing the zeroth output of the node.
@@ -56,7 +55,6 @@ public:
     Node* get_node() const;
     /// \return A `shared_ptr` to the node referred to by this output handle.
     ///
-    /// TODO: Make a plan to deprecate this.
     std::shared_ptr<Node> get_node_shared_ptr() const;
 
     /// \return The index of the output referred to by this output handle.
@@ -84,7 +82,6 @@ public:
     /// \brief Removes a target input from the output referenced by this output handle.
     /// \param target_input The target input to remove.
     ///
-    // TODO(amprocte): Investigate whether this really ought to be public.
     void remove_target_input(const Input<Node>& target_input) const;
 
     /// \brief Replace all users of this value with replacement
@@ -114,7 +111,6 @@ public:
     /// \param node A `shared_ptr` to the node for the output handle.
     /// \param index The index of the output.
     ///
-    /// TODO: Make a plan to deprecate this.
     Output(const std::shared_ptr<const Node>& node, size_t index);
 
     /// \brief Constructs a Output, referencing the zeroth output of the node.
@@ -134,7 +130,6 @@ public:
     const Node* get_node() const;
     /// \return A `shared_ptr` to the node referred to by this output handle.
     ///
-    /// TODO: Make a plan to deprecate this.
     std::shared_ptr<const Node> get_node_shared_ptr() const;
     /// \return The index of the output referred to by this output handle.
     size_t get_index() const;
